Update gradle dependencies. Rename CalculatorAppTest to SystemAppsTest to make more sense. Make failing unit tests pass (especially on emulators), and catch failures on incompatible devices.
